+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ Changes:
   * Improved performance of bulk library pages
   * Changed sample aliquot purpose to be optional (detailed sample)
   * Modified library Tracking List downloads to include index names
+  * Added Fragment Analyser sheet downloads from the Edit Box page
   * Fixed Edit button on Scientific Names list to link to the correct page
   * Fixed Children action on Samples list to work in plain sample mode
   * Fixed Download action on Samples list to work in plain sample mode